requirements.txt
streamlit
openai
requests
python-dotenv
moviepy==1.0.3
edge-tts
pandas
imageio==2.33.1
imageio-ffmpeg>=0.4.9
decorator>=4.0.11
tqdm
pillow
numpy